Adding files to the projects
Makefile is still empty but not for long I hope.
README starts to be something real.
src/archives.html is the templates for the archives page.
src/exclude is for .git/info/exclude
src/install.sh is the main script which will include other when generating
  fugitive.
src/post-receive.sh is still empty for now.